Adding CNNs to Recurrent Neural Network (RNNs) or Long Short Term Memory (LSTM) to improve forecasts
We’ll move away from synthetic data to do a real-world prediction. We will use a dataset from Jason Brownlee, author of the amazing MachineLearningMastery.com site and who has shared lots of datasets at https://github.com/jbrownlee/Datasets. It’s a dataset of daily minimum temperatures in the city of Melbourne, Australia measured from 1981 to 1990.  Your task is to parse the data from a CSV, create a time series and build a prediction model from it.